# Optimize resetting gesture recognizers