This should improve the performance of resetting the gesture recognizers on any View by leveraging a HashSet and reducing casting