@mattleibow 2024-01-19
Is it possible to add a benchmark for this?