@symbiogenesis 2024-01-19
Before:

| Method                | Mean     | Error   | StdDev  | Gen0       | Gen1      | Allocated |
|---------------------- |---------:|--------:|--------:|-----------:|----------:|----------:|
| GetVisualTreeElements | 199.6 ms | 3.72 ms | 5.68 ms | 26000.0000 | 4500.0000 |  234.9 MB |

After:

| Method                | Mean     | Error   | StdDev  | Gen0       | Gen1      | Allocated |
|---------------------- |---------:|--------:|--------:|-----------:|----------:|----------:|
| GetVisualTreeElements | 192.1 ms | 3.57 ms | 6.35 ms | 24500.0000 | 4000.0000 | 221.19 MB |