@symbiogenesis 2024-01-20
Still working on it. Current iteration looks like this:

| Method                         | Mean         | Error      | StdDev     | Gen0      | Gen1     | Gen2    | Allocated  |
|------------------------------- |-------------:|-----------:|-----------:|----------:|---------:|--------:|-----------:|
| AddLotsOfGestureRecognizers    | 11,889.09 us | 261.119 us | 723.560 us | 1093.7500 | 125.0000 | 31.2500 | 9843.75 KB |
| ClearLotsOfGestureRecognizers  |  2,895.50 us |  55.361 us |  63.753 us |  238.2813 | 234.3750 | 97.6563 | 2218.98 KB |
| RemoveLotsOfGestureRecognizers |     55.17 us |   1.037 us |   1.018 us |    0.3052 |   0.2441 |       - |    2.93 KB |
| AddOneGestureRecognizer        |  3,932.09 us |  49.127 us |  43.550 us |  437.5000 | 429.6875 | 15.6250 | 4047.62 KB |
| RemoveOneGestureRecognizer     |  2,373.05 us |  45.308 us |  50.360 us |  292.9688 | 289.0625 |       - | 2719.46 KB |
| ClearOneGestureRecognizer      |  3,896.27 us |  77.675 us | 155.126 us |  562.5000 | 554.6875 |       - | 5172.02 KB |

After:


| Method                         | Mean         | Error      | StdDev       | Gen0      | Gen1     | Gen2    | Allocated  |
|------------------------------- |-------------:|-----------:|-------------:|----------:|---------:|--------:|-----------:|
| AddLotsOfGestureRecognizers    | 13,859.91 us | 585.397 us | 1,716.868 us | 1093.7500 | 125.0000 | 31.2500 | 9843.73 KB |
| ClearLotsOfGestureRecognizers  |    748.43 us |  14.558 us |    18.930 us |  184.5703 | 152.3438 | 75.1953 | 1806.29 KB |
| RemoveLotsOfGestureRecognizers |     51.87 us |   0.961 us |     0.944 us |    0.3052 |   0.2441 |       - |    2.93 KB |
| AddOneGestureRecognizer        |  3,929.22 us |  74.825 us |    76.840 us |  437.5000 | 433.5938 | 15.6250 | 4047.61 KB |
| RemoveOneGestureRecognizer     |  2,473.48 us |  38.957 us |    36.441 us |  292.9688 | 289.0625 |       - | 2719.46 KB |
| ClearOneGestureRecognizer      |  3,809.23 us |  67.035 us |    68.840 us |  546.8750 | 539.0625 |       - | 5078.26 KB |