@symbiogenesis 2024-01-20
Reworked it a bit.

Before:

| Method                         | Mean         | Error      | StdDev     | Gen0      | Gen1     | Gen2    | Allocated  |
|------------------------------- |-------------:|-----------:|-----------:|----------:|---------:|--------:|-----------:|
| AddLotsOfGestureRecognizers    | 11,889.09 us | 261.119 us | 723.560 us | 1093.7500 | 125.0000 | 31.2500 | 9843.75 KB |
| ClearLotsOfGestureRecognizers  |  2,895.50 us |  55.361 us |  63.753 us |  238.2813 | 234.3750 | 97.6563 | 2218.98 KB |
| RemoveLotsOfGestureRecognizers |     55.17 us |   1.037 us |   1.018 us |    0.3052 |   0.2441 |       - |    2.93 KB |
| AddOneGestureRecognizer        |  3,932.09 us |  49.127 us |  43.550 us |  437.5000 | 429.6875 | 15.6250 | 4047.62 KB |
| RemoveOneGestureRecognizer     |  2,373.05 us |  45.308 us |  50.360 us |  292.9688 | 289.0625 |       - | 2719.46 KB |
| ClearOneGestureRecognizer      |  3,896.27 us |  77.675 us | 155.126 us |  562.5000 | 554.6875 |       - | 5172.02 KB |

After:


| Method                         | Mean         | Error      | StdDev       | Gen0      | Gen1     | Gen2    | Allocated  |
|------------------------------- |-------------:|-----------:|-------------:|----------:|---------:|--------:|-----------:|
| AddLotsOfGestureRecognizers    | 12,995.93 us | 501.774 us | 1,471.616 us | 1015.6250 | 125.0000 | 31.2500 | 9154.28 KB |
| ClearLotsOfGestureRecognizers  |    563.02 us |  11.034 us |    10.837 us |  149.4141 | 128.9063 | 52.7344 | 1478.13 KB |
| RemoveLotsOfGestureRecognizers |     50.46 us |   1.003 us |     1.304 us |    0.3052 |   0.2441 |       - |    2.93 KB |
| AddOneGestureRecognizer        |  3,599.49 us |  59.508 us |    52.753 us |  433.5938 | 429.6875 | 15.6250 | 3992.92 KB |
| RemoveOneGestureRecognizer     |  2,364.28 us |  46.301 us |    72.084 us |  292.9688 | 289.0625 |       - | 2719.46 KB |
| ClearOneGestureRecognizer      |  3,651.68 us |  72.749 us |   121.547 us |  542.9688 | 539.0625 |  3.9063 | 5023.57 KB |