@symbiogenesis 2024-01-21
Reworked it a bit.

Allocations are down, and it looks like clearing can be up to six times faster with this PR.

Before:

| Method                         | Mean         | Error      | StdDev     | Gen0      | Gen1     | Gen2    | Allocated  |
|------------------------------- |-------------:|-----------:|-----------:|----------:|---------:|--------:|-----------:|
| AddLotsOfGestureRecognizers    | 11,889.09 us | 261.119 us | 723.560 us | 1093.7500 | 125.0000 | 31.2500 | 9843.75 KB |
| ClearLotsOfGestureRecognizers  |  2,895.50 us |  55.361 us |  63.753 us |  238.2813 | 234.3750 | 97.6563 | 2218.98 KB |
| RemoveLotsOfGestureRecognizers |     55.17 us |   1.037 us |   1.018 us |    0.3052 |   0.2441 |       - |    2.93 KB |
| AddOneGestureRecognizer        |  3,932.09 us |  49.127 us |  43.550 us |  437.5000 | 429.6875 | 15.6250 | 4047.62 KB |
| RemoveOneGestureRecognizer     |  2,373.05 us |  45.308 us |  50.360 us |  292.9688 | 289.0625 |       - | 2719.46 KB |
| ClearOneGestureRecognizer      |  3,896.27 us |  77.675 us | 155.126 us |  562.5000 | 554.6875 |       - | 5172.02 KB |

After:


| Method                         | Mean         | Error      | StdDev     | Median       | Gen0      | Gen1     | Gen2    | Allocated  |
|------------------------------- |-------------:|-----------:|-----------:|-------------:|----------:|---------:|--------:|-----------:|
| AddLotsOfGestureRecognizers    | 10,579.72 us | 199.913 us | 442.994 us | 10,593.28 us | 1000.0000 | 125.0000 | 31.2500 | 9154.29 KB |
| ClearLotsOfGestureRecognizers  |    542.30 us |  10.537 us |  13.326 us |    540.67 us |  149.4141 | 128.9063 | 52.7344 | 1478.13 KB |
| RemoveLotsOfGestureRecognizers |     52.00 us |   0.846 us |   1.838 us |     51.57 us |    0.3052 |   0.2441 |       - |    2.93 KB |
| AddOneGestureRecognizer        |  3,607.38 us |  63.466 us |  59.366 us |  3,612.58 us |  433.5938 | 429.6875 | 15.6250 | 3992.92 KB |
| RemoveOneGestureRecognizer     |  2,702.21 us | 224.002 us | 660.474 us |  2,344.16 us |  292.9688 | 289.0625 |       - | 2719.46 KB |
| ClearOneGestureRecognizer      |  3,647.28 us |  71.943 us | 129.729 us |  3,693.72 us |  546.8750 | 542.9688 |  7.8125 | 5023.58 KB |