@symbiogenesis 2024-01-21
Although these changes helped a bit, I may close this one and move the benchmark file to my new PR https://github.com/dotnet/maui/pull/20036 

Then, I can make all the ObservableCollections of GestureRecognizers into ObservableRangeCollections and do a very simple AddRange() RemoveRange() and ReplaceRange() operations that will dramatically reduce CollectionChanged events.