@jonathanpeppers 2024-01-22
@symbiogenesis I just wanted to ask how you decide which method(s) to focus on for some of your recent PRs.

* Are you profiling a real-world MAUI app? (with either VS, PerfView, or `dotnet-trace`)
* Then look at methods that show time spent (wasted?) in a real app

I worry you might be optimizing some methods that aren't as impactful as others. If you see it in a real app, when profiling, can you share a screenshot of how much time it was taking?